+++ packages/patrol/android/src/main/kotlin/pl/leancode/patrol/contracts/Contracts.kt
@@ -125,6 +125,18 @@ class Contracts {
     val appId: String
   )
 
+  data class GetNativeUITreeRequest (
+    val iosInstalledApps: List<String>? = null
+  ){
+    fun hasIosInstalledApps(): Boolean {
+      return iosInstalledApps != null
+    }
+  }
+
+  data class GetNativeUITreeRespone (
+    val roots: List<NativeView>
+  )
+
   data class NativeView (
     val className: String? = null,
     val text: String? = null,