+++ packages/patrol/ios/Classes/AutomatorServer/Contracts.swift
@@ -83,6 +83,14 @@ struct GetNativeViewsRequest: Codable {
   var appId: String
 }
 
+struct GetNativeUITreeRequest: Codable {
+  var iosInstalledApps: [String]?
+}
+
+struct GetNativeUITreeRespone: Codable {
+  var roots: [NativeView]
+}
+
 struct NativeView: Codable {
   var className: String?
   var text: String?