+++ packages/patrol/lib/src/native/contracts/contracts.dart
@@ -259,6 +259,44 @@ class GetNativeViewsRequest with EquatableMixin {
 }
 
 @JsonSerializable()
+class GetNativeUITreeRequest with EquatableMixin {
+  GetNativeUITreeRequest({
+    this.iosInstalledApps,
+  });
+
+  factory GetNativeUITreeRequest.fromJson(Map<String, dynamic> json) =>
+      _$GetNativeUITreeRequestFromJson(json);
+
+  final List<String>? iosInstalledApps;
+
+  Map<String, dynamic> toJson() => _$GetNativeUITreeRequestToJson(this);
+
+  @override
+  List<Object?> get props => [
+        iosInstalledApps,
+      ];
+}
+
+@JsonSerializable()
+class GetNativeUITreeRespone with EquatableMixin {
+  GetNativeUITreeRespone({
+    required this.roots,
+  });
+
+  factory GetNativeUITreeRespone.fromJson(Map<String, dynamic> json) =>
+      _$GetNativeUITreeResponeFromJson(json);
+
+  final List<NativeView> roots;
+
+  Map<String, dynamic> toJson() => _$GetNativeUITreeResponeToJson(this);
+
+  @override
+  List<Object?> get props => [
+        roots,
+      ];
+}
+
+@JsonSerializable()
 class NativeView with EquatableMixin {
   NativeView({
     this.className,