+++ packages/patrol/lib/src/native/contracts/contracts.g.dart
@@ -141,6 +141,34 @@ Map<String, dynamic> _$GetNativeViewsRequestToJson(
       'appId': instance.appId,
     };
 
+GetNativeUITreeRequest _$GetNativeUITreeRequestFromJson(
+        Map<String, dynamic> json) =>
+    GetNativeUITreeRequest(
+      iosInstalledApps: (json['iosInstalledApps'] as List<dynamic>?)
+          ?.map((e) => e as String)
+          .toList(),
+    );
+
+Map<String, dynamic> _$GetNativeUITreeRequestToJson(
+        GetNativeUITreeRequest instance) =>
+    <String, dynamic>{
+      'iosInstalledApps': instance.iosInstalledApps,
+    };
+
+GetNativeUITreeRespone _$GetNativeUITreeResponeFromJson(
+        Map<String, dynamic> json) =>
+    GetNativeUITreeRespone(
+      roots: (json['roots'] as List<dynamic>)
+          .map((e) => NativeView.fromJson(e as Map<String, dynamic>))
+          .toList(),
+    );
+
+Map<String, dynamic> _$GetNativeUITreeResponeToJson(
+        GetNativeUITreeRespone instance) =>
+    <String, dynamic>{
+      'roots': instance.roots,
+    };
+
 NativeView _$NativeViewFromJson(Map<String, dynamic> json) => NativeView(
       className: json['className'] as String?,
       text: json['text'] as String?,