+++ packages/patrol/lib/src/native/contracts/native_automator_client.dart
@@ -96,6 +96,16 @@ class NativeAutomatorClient {
     );
   }
 
+  Future<GetNativeUITreeRespone> getNativeUITree(
+    GetNativeUITreeRequest request,
+  ) async {
+    final json = await _sendRequest(
+      'getNativeUITree',
+      request.toJson(),
+    );
+    return GetNativeUITreeRespone.fromJson(json);
+  }
+
   Future<GetNativeViewsResponse> getNativeViews(
     GetNativeViewsRequest request,
   ) async {