+++ packages/patrol_cli/lib/src/commands/develop.dart
@@ -49,6 +49,12 @@ class DevelopCommand extends PatrolCommand {
 
     usesAndroidOptions();
     usesIOSOptions();
+
+    argParser.addFlag(
+      'open-devtools',
+      help: 'Automatically open Patrol extension in DevTools when ready.',
+      defaultsTo: true,
+    );
   }
 
   final DeviceFinder _deviceFinder;
@@ -112,6 +118,12 @@ class DevelopCommand extends PatrolCommand {
     final displayLabel = boolArg('label');
     final uninstall = boolArg('uninstall');
 
+    String? iOSInstalledAppsEnvVariable;
+    if (device.targetPlatform == TargetPlatform.iOS) {
+      iOSInstalledAppsEnvVariable =
+          await _iosTestBackend.getInstalledAppsEnvVariable(device.id);
+    }
+
     final customDartDefines = {
       ..._dartDefinesReader.fromFile(),
       ..._dartDefinesReader.fromCli(args: stringsArg('dart-define')),
@@ -125,7 +137,10 @@ class DevelopCommand extends PatrolCommand {
       'INTEGRATION_TEST_SHOULD_REPORT_RESULTS_TO_NATIVE': 'false',
       'PATROL_TEST_LABEL_ENABLED': displayLabel.toString(),
       // develop-specific
-      ...{'PATROL_HOT_RESTART': 'true'},
+      ...{
+        'PATROL_HOT_RESTART': 'true',
+        'PATROL_IOS_INSTALLED_APPS': iOSInstalledAppsEnvVariable,
+      },
     }.withNullsRemoved();
 
     final dartDefines = {...customDartDefines, ...internalDartDefines};
@@ -170,6 +185,7 @@ class DevelopCommand extends PatrolCommand {
       iosOpts,
       uninstall: uninstall,
       device: device,
+      openDevtools: boolArg('open-devtools'),
     );
 
     return 0; // for now, all exit codes are 0
@@ -245,6 +261,7 @@ class DevelopCommand extends PatrolCommand {
     IOSAppOptions iosOpts, {
     required bool uninstall,
     required Device device,
+    required bool openDevtools,
   }) async {
     Future<void> Function() action;
     Future<void> Function()? finalizer;
@@ -282,6 +299,7 @@ class DevelopCommand extends PatrolCommand {
         target: flutterOpts.target,
         appId: appId,
         dartDefines: flutterOpts.dartDefines,
+        openDevtools: openDevtools,
       );
 
       await future;