+++ packages/patrol_cli/lib/src/runner/patrol_command_runner.dart
@@ -158,6 +158,7 @@ class PatrolCommandRunner extends CompletionCommandRunner<int> {
         flutterTool: FlutterTool(
           stdin: stdin,
           processManager: _processManager,
+          platform: _platform,
           parentDisposeScope: _disposeScope,
           logger: _logger,
         ),