+++ schema.dart
@@ -69,6 +69,14 @@ class GetNativeViewsRequest {
   late String appId;
 }
 
+class GetNativeUITreeRequest {
+  List<String>? iosInstalledApps;
+}
+
+class GetNativeUITreeRespone {
+  late List<NativeView> roots;
+}
+
 class NativeView {
   String? className;
   String? text;
@@ -178,6 +186,7 @@ abstract class NativeAutomator<IOSServer, AndroidServer, DartClient> {
   void openQuickSettings(OpenQuickSettingsRequest request);
 
 // general UI interaction
+  GetNativeUITreeRespone getNativeUITree(GetNativeUITreeRequest request);
   GetNativeViewsResponse getNativeViews(GetNativeViewsRequest request);
   void tap(TapRequest request);
   void doubleTap(TapRequest request);